+++ js/data.js
@@ -1,4 +1,4 @@
-import {getRandomArrayElement} from './util.js';
+import { getRandomPositiveInteger, getRandomArrayElement, createIdGenerator } from './util.js';
 
 const POSTED_PICTURES = 25;
 const LIKES_MIN = 15;
@@ -11,14 +11,6 @@ const MESSAGES = [
 ];
 const DESCRIPTIONS = ['Это мой лучший кадр!', 'Я будущий популярный блогер!'];
 const NAMES = ['Оксана', 'Матвей', 'Виктория', 'Тимофей', 'Юрий'];
-
-const createIdGenerator = () => {
-  let lastGeneratedId = 0;
-  return () => {
-    lastGeneratedId += 1;
-    return lastGeneratedId;
-  };
-};
 
 const generateCommentId = createIdGenerator();
 
@@ -50,6 +42,4 @@ const getPictures = () =>
     createPicture(pictureIndex + 1)
   );
 
-getPictures();
-
-export {createIdGenerator, createMessage, createComment, createPicture, getPictures};
+export { getPictures };
