+++ js/main.js
@@ -1,1 +1,3 @@
-import {createIdGenerator, createMessage, createComment, createPicture, getPictures} from './data.js';
+import { getPictures } from './data.js';
+
+getPictures();
